requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==2.3.3
python-dotenv==1.0.0
cachetools==5.3.1
fake-useragent==1.2.1
flask-cors==4.0.0
gunicorn==21.2.0
